requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# serverstats.py: 1,4
# serverstatus_client.py: 1,5
discord.py == 1.0.0a1540+gbd46708

# serverstats.py: 8
nvidia_ml_py3 == 7.352.0

# serverstats.py: 5
psutil == 5.4.0

# serverstats.py: 3
python_dateutil == 2.6.1
